+++ CHANGELOG.md
@@ -41,6 +41,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 * [#598](https://github.com/babylonlabs-io/finality-provider/pull/598) chore: recover-rand-proof cosmos
 
+### Bug fixes
+
+* [#602](https://github.com/babylonlabs-io/finality-provider/pull/602) fix: use latest height when for ctxsigning
+
 ## v2.0.0-rc.0
 
 ### Bug Fixes